+++ src/NoteWorkspace.ts
@@ -1,21 +1,12 @@
 import * as vscode from 'vscode';
-import { existsSync, readFileSync } from 'fs';
+import { existsSync } from 'node:fs';
 import { join } from 'path';
+
 
 export class NoteWorkspace {
     static _rxFileExtensions = '\\.(md|markdown|mdx|fountain|txt)$';
 
-    static newNoteContent(title: string, noteDirectory: string, useTemplate:boolean) {
-        const path = join(noteDirectory, '_NoteTemplate.md');
-        if(useTemplate){
-            const fileContents = readFileSync(path).toString();
-            return `# ${title}\n\n${fileContents}`;
-
-        }
-        return `# ${title}\n\n`;
-    }
-
-    static newNoteUsingTemplate(){
+    static newNoteUsingTemplate() {
         NoteWorkspace.newNote(true);
     }
 
@@ -44,6 +35,53 @@ export class NoteWorkspace {
             });
     }
 
+    static sortTwoLevelList() {
+        const editor = vscode.window.activeTextEditor;
+    
+        if (!editor) {
+          // Abort: no active editor;
+          return;
+        }
+        const { selection } = editor;
+        const text = editor.document.getText(selection);
+        const selectionRange = new vscode.Range(selection.start, selection.end);
+    
+        if (text === '') {
+          vscode.window.showErrorMessage('Error creating note from selection: selection is empty.');
+          return;
+        }
+
+        let sortedList = NoteWorkspace.sortList(text);
+
+        // Replace the selected content in the origin file with a wiki-link to the new file
+        const edit = new vscode.WorkspaceEdit();
+
+
+        edit.replace(
+          editor.document.uri,
+          selectionRange,
+          sortedList
+        );
+
+        vscode.workspace.applyEdit(edit);
+    } 
+    
+    static sortList(text: string): string {
+        const regex = /^-.+\n(\s+-.+\n*)*/gm;
+
+        // Make sure selection contains only a list
+        let foundIndex = text.search(regex);
+        let textArray = text.match(regex);
+    
+        if (foundIndex === 0 && textArray) {
+            let trimmed = textArray.map(element => element.trim()).sort();
+            return trimmed.join("\n");
+        }
+    
+        console.log("No match found, only select a list");
+        return text;
+    }
+
     static getWorkspaceDirectory() {
         let dir = '';
         if (vscode.workspace.workspaceFolders !== undefined) {
@@ -52,8 +90,26 @@ export class NoteWorkspace {
         return dir;
     }
 
+    static async newNoteContent(title: string, noteDirectory: string, useTemplate: boolean) {
+        const path = join(noteDirectory, '_NoteTemplate.md');
+
+        if (useTemplate) {
+
+            const uri = vscode.Uri.file(path);
+            try {
+                const fileContent = await vscode.workspace.fs.readFile(uri);
+                const fileContentString = Buffer.from(fileContent).toString('utf8');
+                return `# ${title}\n\n${fileContentString}`;
+            } catch (error) {
+                vscode.window.showErrorMessage(`Error reading file: ${error}`);
+            }
+
+            return `# ${title}\n\n`;
+        }
+    }
+
     static async createNewNoteFile(noteDirectory: string, useTemplate: boolean) {
-   
+
         const d = (new Date().toISOString().match(/(\d{4}-\d{2}-\d{2})/) || '')[0]; // "2020-08-25"
         const filename = this.noteFileNameFromTitle(d);
         const filepath = join(noteDirectory, filename);
@@ -61,7 +117,7 @@ export class NoteWorkspace {
         const fileAlreadyExists = existsSync(filepath);
         if (!fileAlreadyExists) {
             // create the file if it does not exist
-            const contents = this.newNoteContent(d, noteDirectory, useTemplate);
+            const contents = await this.newNoteContent(d, noteDirectory, useTemplate);
             const edit = new vscode.WorkspaceEdit();
             const fileUri = vscode.Uri.file(filepath.toString());
             edit.createFile(fileUri);