+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "VsCodeExtensionCoding" extension will be documented 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.1]
 
 - Initial release
+
+## [1.0.1]
+
+- Added try catch on reading template
+
+## [1.0.2]
+
+- Using vscode.workspace.fs to read file